fix: improve validation steps
